Update webpacker gem and webpack-dev-server
In order to address CVE-2018-14732, it was necessary to update `webpack-dev-server` to `>= 3.1.11`. Since the latest published version of `webpacker` was forcing `webpack-dev-server` to version `2.x`, it was necessary to update `webpacker` gem to latest master branch, that already had a fix.

Also, I updated jest configuration, removing from `packages.json` file and placing it on `jest.config.js`.