[SHAME] This commit fixes a code that would never work.
The excess of confidence and hurry on delivering code makes a developer push code without proper testing. Fortunately, such code would never reach production because automated test suite was good enough to capture it before hand.

This is a constant reminder to future me that, regardless of experience and knowledge in a language or domain, I won't make such mistakes.